+++ tests/unit/PaymentMethodTest.php
@@ -35,8 +35,11 @@ class PaymentMethodTest extends Setup
                 'verificationAmount',
                 ['paypal' => [
                     'payee_email',
+                    'payeeEmail',
                     'order_id',
+                    'orderId',
                     'custom_field',
+                    'customField',
                     'description',
                     'amount',
                 ]],